pyproject.toml
[project]
name = "huntstand-exporter"
version = "0.2.0"
description = "CLI tool to export HuntStand hunt area membership data (members, invites, requests) to CSV/JSON."
authors = [ { name = "Anthony Pinto" } ]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9"
keywords = ["huntstand", "export", "csv", "membership", "scraper"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Utilities"
]
dependencies = [
  "requests>=2.25.0",
  "certifi>=2020.6.20",
  "python-dotenv>=0.19.0",
  "urllib3>=1.26.0"
]

[project.urls]
Homepage = "https://github.com/anthnyajp/HuntStand"
Repository = "https://github.com/anthnyajp/HuntStand"
Issues = "https://github.com/anthnyajp/HuntStand/issues"
Changelog = "https://github.com/anthnyajp/HuntStand/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "types-requests>=2.31.0",
  "build>=1.0.0"
]

[project.scripts]
huntstand-exporter = "huntstand_exporter.exporter:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
